[UPDATE] Delete corregido y triggers

+++ src/main/java/utez/edu/mx/practica1/model/person/DaoPerson.java
@@ -51,6 +51,7 @@ public class DaoPerson {
         JSONObject newPerson = new JSONObject();
         Boolean flag =false;
         try{
+            rs =null;
             con = ConnectionDB.getConnection();
             pstm = con.prepareStatement("INSERT INTO persona(nombre, a_paterno, a_materno, edad, sexo, telefono, direccion, " +
                     "fecha_nacimiento, estado_civil, correo, contrasena,estado) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?,?)", PreparedStatement.RETURN_GENERATED_KEYS);
@@ -131,6 +132,7 @@ public class DaoPerson {
     public BeanPerson getPersonByLogin(String correo,String contrasena){
         BeanPerson personLog = null;
         try{
+            rs = null;
             con = ConnectionDB.getConnection();
             pstm = con.prepareStatement("SELECT * FROM persona WHERE correo = ? AND contrasena = ?;");
             pstm.setString(1,correo);
@@ -166,8 +168,10 @@ public class DaoPerson {
         BeanPerson person = null;
         JSONObject selPer = new JSONObject();
         try{
+            rs = null;
             con = ConnectionDB.getConnection();
             pstm = con.prepareStatement("SELECT * FROM persona");
+
             rs = pstm.executeQuery();
             while (rs.next()) {
                 person = new BeanPerson();
@@ -200,7 +204,7 @@ public class DaoPerson {
         }catch (Exception e) {
             System.out.println("DP_ERR_01" + e.getMessage());
         }finally {
-            ConnectionDB.closeConnections(con,pstm,rs);
+            ConnectionDB.closeConnections(con,pstm);
         }
         return listPersons;
     }
@@ -208,10 +212,11 @@ public class DaoPerson {
 
     
     // EncontrarPorId
-    public BeanPerson findById(Long id){
+    public BeanPerson findById(Long id,Long idSeP){
         BeanPerson person = null;
         JSONObject selPer = new JSONObject();
         try {
+            rs = null;
             con = ConnectionDB.getConnection();
             pstm = con.prepareStatement("SELECT * FROM persona WHERE id = ?");
             pstm.setLong(1,id);
@@ -233,7 +238,7 @@ public class DaoPerson {
                 person.setEstado(rs.getBoolean("estado"));
                 person.setContrasena(rs.getString("contrasena"));
 
-                Long idPer = rs.getLong("id");
+
                 selPer.put("id",rs.getLong("id"));
                 selPer.put("nombre",rs.getString("nombre"));
                 selPer.put("aPaterno",rs.getString("a_Paterno"));
@@ -243,7 +248,7 @@ public class DaoPerson {
                 //QUERY PARA GENERAR LA OPERACIÓN
                 pstm = con.prepareStatement("INSERT INTO json_persona(datos,id_persona,tipoaccion) values(JSON_INSERT('{}','$.json',?),?,?);");
                 pstm.setString(1,selPer.toString());
-                pstm.setLong(2,idPer);
+                pstm.setLong(2,idSeP);
                 pstm.setString(3,"findbyid");
                 pstm.executeUpdate();
 
@@ -258,7 +263,7 @@ public class DaoPerson {
     }
 
     // Create
-    public boolean create(BeanPerson person){
+    public boolean create(BeanPerson person,Long idSesiP){
         boolean flag = false;
         JSONObject newPerson = new JSONObject();
         try{
@@ -293,33 +298,24 @@ public class DaoPerson {
             newPerson.put("estado",person.isEstado());
 
             if(pstm.executeUpdate() == 1){
-                ResultSet lastInsert = pstm.getGeneratedKeys();
-                lastInsert.next();
-                Long lastId = lastInsert.getLong(1);
-                pstm = con.prepareStatement("SELECT * FROM persona WHERE id = ?");
-                pstm.setLong(1, lastId);
-                rs = pstm.executeQuery();
-                lastInsert.close();
-                flag = true;
-                if(rs.next()){
+
                     //QUERY PARA GENERAR LA OPERACIÓN
                      pstm = con.prepareStatement("INSERT INTO json_persona(datos,id_persona,tipoaccion) values(JSON_INSERT('{}','$.json',?),?,?);");
                      pstm.setString(1, newPerson.toString());
-                     pstm.setLong(2,lastId);
+                     pstm.setLong(2,idSesiP);
                      pstm.setString(3,"create");
-                     pstm.executeUpdate();
-                }
+                    flag =  pstm.executeUpdate() == 1;
             }
         }catch (SQLException e){
             System.out.println("DP_ERR_003" + e.getMessage());
         }finally {
-            ConnectionDB.closeConnections(con,pstm,rs);
+            ConnectionDB.closeConnections(con,pstm);
         }
         return flag;
     }
 
     // Update
-    public boolean update(BeanPerson person){
+    public boolean update(BeanPerson person,Long idSesiP){
         boolean flag = false;
         JSONObject newPerson = new JSONObject();
         try{
@@ -359,7 +355,7 @@ public class DaoPerson {
                     pstm = con.prepareStatement("UPDATE json_persona SET datos = JSON_SET('{}','$.json',?),tipoaccion = ? WHERE id_persona = ?;");
                     pstm.setString(1,newPerson.toString());
                     pstm.setString(2,"update");
-                pstm.setLong(3,person.getId());
+                pstm.setLong(3,idSesiP);
                     pstm.executeUpdate();
                 flag = true;
             }
@@ -386,43 +382,63 @@ public class DaoPerson {
     }
 
     // Delete
-    public boolean delete(Long id){
-        boolean flag=false;
-        try{
-            con = ConnectionDB.getConnection();
-            pstm = con.prepareCall("DELETE FROM json_persona WHERE id_persona = ?;");
-            pstm.setLong(1,id);
-            if(pstm.executeUpdate() == 1){
-                    //QUERY PARA GENERAR LA OPERACIÓN
-                pstm = con.prepareCall("DELETE FROM persona WHERE id=?");
-                pstm.setLong(1,id);
-                   if( pstm.executeUpdate() == 1){
-                       flag = true;
-                   }
+    public boolean delete(Long id,Long idSesiP){
+        boolean flag = false;
+        if(id == idSesiP){
+            flag = false;
+        }else{
+            System.out.println("Eliminando");
+            try{
+                System.out.println("Apunto de");
+                String sql = "SET @idSesiP = ?";
+                con = ConnectionDB.getConnection();
+                pstm = con.prepareStatement(sql);
+                pstm.setLong(1,idSesiP);
+                pstm.executeUpdate();
+                    pstm = con.prepareCall("DELETE FROM persona WHERE id=?;");
+                    pstm.setLong(1,id);
+                    if(pstm.executeUpdate() == 1){
+                        System.out.println("Se elimino acciones de esa persona");
+                      //  pstm.close();
+                       pstm = con.prepareCall("DELETE FROM json_persona WHERE id_persona=?;");
+                       pstm.setLong(1,id);
+                       if(pstm.executeUpdate() == 1){
+                           System.out.println("Se elimino a la persona");
+                           flag = true;
+                       }
+                    }
 
+            }catch(SQLException e){
+                System.out.println("DP_ERR_005" + e.getMessage());
             }
-
-
-        }catch(SQLException e){
-            System.out.println("DP_ERR_005" + e.getMessage());
-        }finally {
-            ConnectionDB.closeConnections(con,pstm);
         }
+
         return flag;
     }
 
-    public List<BeanTransaction> getTransactions () {
+    public List<BeanTransaction> getTransactions (Long idSesiP) {
         List<BeanTransaction> transactions = new ArrayList<>();
         BeanTransaction opera = null;
         try{
+            rs = null;
             con = ConnectionDB.getConnection();
-            pstm = con.prepareStatement("SELECT op.id,CONCAT(P.nombre, ' ', p.a_paterno, ' ', p.a_materno ) as nombre,op.fecha as fechamovimiento,t.nombre as tipo,op.dato_viejo,op.dato_nuevo FROM operacion op join json_persona jp on op.id_jsonpersona = jp.id\n" +
-                    "join persona p on jp.id_persona = p.id\n" +
-                    "join tipo_operacion t on op.id_tipo_operacion = t.id;");
+            pstm = con.prepareStatement("SELECT op.id,CONCAT(P.nombre, ' ', p.a_paterno, ' ', p.a_materno ) as nombre,op.fecha as fechamovimiento,t.nombre as tipo,JSON_EXTRACT(op.dato_viejo,'$.json') as dato_viejo,op.dato_nuevo\n" +
+                    "FROM operacion op join json_persona jp on op.id_jsonpersona = jp.id\n" +
+                    "                    join persona p on jp.id_persona = p.id\n" +
+                    "                    join tipo_operacion t on op.id_tipo_operacion = t.id\n" +
+                    "                    where p.id = ?\n" +
+                    "union all\n" +
+                    "SELECT op.id,CONCAT(P.nombre, ' ', p.a_paterno, ' ', p.a_materno ) as nombre,op.fecha as fechamovimiento,t.nombre as tipo,op.dato_viejo as dato_viejo,op.dato_nuevo\n" +
+                    "FROM operacion op\n" +
+                    "                    join persona p on op.idPersonaD = p.id\n" +
+                    "                    join tipo_operacion t on op.id_tipo_operacion = t.id\n" +
+                    "                    where idPersonaD = ?;");
+            pstm.setLong(1,idSesiP);
+            pstm.setLong(2,idSesiP);
             rs = pstm.executeQuery();
 
             while(rs.next()){
-                    opera = new BeanTransaction();
+                opera = new BeanTransaction();
                 opera.setId(rs.getInt("id"));
                 opera.setNombrePersona(rs.getString("nombre"));
                 opera.setFechaMovimiento(rs.getString("fechaMovimiento"));
@@ -433,9 +449,9 @@ public class DaoPerson {
             }
 
         }catch (SQLException e){
-            System.out.println("DP_ERR_00TS");
+            System.out.println("DP_ERR_00TS" + e.getMessage());
         }finally {
-            ConnectionDB.closeConnections(con,pstm,rs);
+            ConnectionDB.closeConnections(con,pstm);
         }
         return transactions;
     }
@@ -443,11 +459,18 @@ public class DaoPerson {
     public BeanTransaction transactionById(Long id){
         BeanTransaction opera= null;
         try {
+            rs = null;
             con = ConnectionDB.getConnection();
-            pstm = con.prepareStatement("SELECT op.id,CONCAT(P.nombre, ' ', p.a_paterno, ' ', p.a_materno ) as nombre,op.fecha as fechamovimiento,t.nombre as tipo,op.dato_viejo,op.dato_nuevo FROM operacion op join json_persona jp on op.id_jsonpersona = jp.id\n" +
-                    "join persona p on jp.id_persona = p.id\n" +
-                    "join tipo_operacion t on op.id_tipo_operacion = t.id WHERE op.id = ?;");
+            pstm = con.prepareStatement("SELECT op.id,CONCAT(P.nombre, ' ', p.a_paterno, ' ', p.a_materno ) as nombre,op.fecha as fechamovimiento,t.nombre as tipo,op.dato_viejo,op.dato_nuevo\n" +
+                    "FROM operacion op join json_persona jp on op.id_jsonpersona = jp.id\n" +
+                    "                    join persona p on jp.id_persona = p.id\n" +
+                    "                    join tipo_operacion t on op.id_tipo_operacion = t.id WHERE op.id = ?\n" +
+                    "union all\n" +
+                    "SELECT op.id,CONCAT(P.nombre, ' ', p.a_paterno, ' ', p.a_materno ) as nombre,op.fecha as fechamovimiento,t.nombre as tipo,op.dato_viejo,op.dato_nuevo\n" +
+                    "FROM operacion op join persona p on op.idPersonaD = p.id\n" +
+                    "                    join tipo_operacion t on op.id_tipo_operacion = t.id WHERE op.id = ?;");
             pstm.setLong(1,id);
+            pstm.setLong(2,id);
             rs = pstm.executeQuery();
             if(rs.next()){
                 opera = new BeanTransaction();
